% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bounds_of.R
\name{calc_bounds_of}
\alias{calc_bounds_of}
\title{Title}
\usage{
calc_bounds_of(
  n_analyses = 3,
  alpha = 0.05,
  n_patients = c(20, 40, 60),
  sided = "one.sided"
)
}
\arguments{
\item{sided}{}
}
\description{
Title
}
