% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bounds_pocock.R
\name{calc_bounds_pocock}
\alias{calc_bounds_pocock}
\title{Calculate Pocock bounds for a group sequential clinical trial}
\usage{
calc_bounds_pocock(
  n_analyses = 3,
  alpha = 0.05,
  n_patients = c(20, 40, 60),
  sided = "one.sided"
)
}
\arguments{
\item{n_analyses}{Integer indicating the number of interim analyses
that will be performed in the group sequential design.}

\item{alpha}{Type I error rate to be targeted when calculating the bounds.}

\item{n_patients}{Vector of integers indicating the sample size (per group)
at each of the interim analyses. Note that the length of this vector must
correspond to the number of analyses \code{n_analyses}.}

\item{sided}{A string "one.sided" or "two.sided" indicating whether the
function returns one- or two-sided bounds. In one-sided bounds, the boundary
values are equal at the last analysis.}
}
\value{
A list of length three that contains the final upper and lower
bounds calculated and the last simulation run. The simulation output is from
\code{gsd_simulations} and shows the probability of stopping for futility and
efficacy under both the null and alternative hypotheses specified. It also
calculates the type I error (\eqn{\alpha}), the power (\eqn{1 - \beta}) where
\eqn{\beta} is the type II error, the expected sample size, and the variance
of the sample size.
}
\description{
\code{calc_bounds_pocock} returns one- or two-sided boundary values for a group
sequential clinical trial design at a specified \eqn{\alpha} and sample
size using interval bisection.
}
\examples{
calc_bounds_pocock(n_analyses = 3,
                   alpha = 0.05,
                   n_patients = c(20, 40, 60),
                   sided = "one.sided")
}
\seealso{
\code{\link[=gsd_simulations]{gsd_simulations()}} for details on the simulation return.
}
