% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bounds_triangular.R
\name{calc_bounds_triangular}
\alias{calc_bounds_triangular}
\title{Title}
\usage{
calc_bounds_triangular(n_analyses = 3, alpha = 0.05, delta = 0.5)
}
\arguments{
\item{delta}{}
}
\description{
Title
}
