% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd_simulations.R
\name{gsd_simulations}
\alias{gsd_simulations}
\title{Simulate a group sequential design for a clinical trial}
\usage{
gsd_simulations(
  n_analyses = 3,
  upper_bounds = c(2.5, 2, 1.5),
  lower_bounds = c(0, 0.75, 1.5),
  n_patients = c(20, 40, 60),
  null_hypothesis = 0,
  alt_hypothesis = 0.5,
  variance = 1
)
}
\arguments{
\item{n_analyses}{Integer indicating the number of interim analyses
that will be performed in the group sequential design.}

\item{upper_bounds}{Value of the upper bounds to use for group sequential
design simulation}

\item{lower_bounds}{Value of the lower bounds to use for group sequential
design simulation}

\item{n_patients}{Vector of integers indicating the sample size (per group)
at each of the interim analyses. Note that the length of this vector must
correspond to the number of analyses \code{n_analyses}.}

\item{null_hypothesis}{The mean difference between groups under the null
hypothesis (default is 0).}

\item{alt_hypothesis}{The mean difference between groups under the
alternative hypothesis (default is 0).}

\item{variance}{The assumed variance in the mean values of the null and
alternative mean differences.}
}
\value{
A list of length (\eqn{n} + 4) where \eqn{n} is the number of analyses
specified by \code{n_analyses}.
\itemize{
\item For each analysis from 1 to \eqn{n} (\code{analysis_1} to \code{analysis_n}) 4
stopping probabilities are returned, the probability of stopping:
\enumerate{
\item for futility under the null.
\item for efficacy under the null.
\item for futility under the alternative.
\item for efficacy under the alternative.
}
\item The expected sample size (ESS).
\item The variance in sample size (VSS).
\item The type I error for the design, (\eqn{alpha}).
\item The power, \eqn{1 - \beta}, where \eqn{\beta} is
the type II error.
}
}
\description{
\code{gsd_simulations} simulates a group sequential design for a clinical trial.
It uses given upper and lower bounds, an assumed null and alternative
hypotheses for a mean difference between groups, an assumed variance for
these mean differences, the number of interim analyses, and sample sizes
(per group) to calculate the probabilities of stopping for futility or
efficacy under the null and alternative hypotheses at each analysis. It also
calculates the expected sample size, the variance of the sample size, and the
type I error (\eqn{\alpha}) and power (\eqn{1 - \beta}) where \eqn{\beta} is
the type II error.
}
\details{
Below is a mathematical description of how each of the values are
calculated.
\subsection{Stopping probabilities}{

Assuming a normally distributed end point, responses for patients on control
treatment is \eqn{Y_{0i} \sim \mathcal{N}(\mu_0, \sigma^2)} and for patients
on experimental treatment is \eqn{Y_{1i} \sim \mathcal{N}(\mu_1, \sigma^2)}.

Assume \eqn{\sigma} is known and the parameter of interest is \eqn{\theta = \mu_1 = \mu_0}.

At analysis \eqn{j}, MLE \eqn{\hat{\theta}_j} is normally distributed with
mean \eqn{\theta} and variance \eqn{\frac{2\sigma^2}{n_j}} and the Wald test
statistic is \eqn{Z_j = \bar{Y}^{(j)}_1 -\bar{Y}^{(j)}_0}. The joint
distribution of each test statistic is multivariate normal.

The following equation is used to calculate the probability of stopping for
futility \strong{under the null} at Analysis 2. It is the lower tail of the
multivariate normal distribution up to the second lower bound (the selected
critical value).

\deqn{

 \int_{l_1}^{u_1}\int_{-\infty}^{l_2} \phi_2 \left( \begin{bmatrix}
 y_1 & y_2
 \end{bmatrix}, \begin{bmatrix}
 \theta \sqrt{\frac{n_1}{2\sigma^2}} & \theta \sqrt{\frac{n_2}{2\sigma^2}}
 \end{bmatrix}, \begin{bmatrix} 1 & \sqrt{\frac{n_1}{n_2}} \\ \sqrt{\frac{n_1}{n_2}} & 1 \end{bmatrix} \right) dy_2\,dy_1

}
}

\subsection{Expected sample size}{

The expected sample size is calculated across a set of possible true treatment
effects \eqn{\boldsymbol{\delta} = \{\delta_1, \delta_2, \dots, \delta_n \}} as
a function of the design elements (1) number of analyses \eqn{k = \{1, 2, \dots, K\}},
(2) number of patients at each analysis \eqn{\mathbf{n} = \{n_1, n_2, \dots, n_K\}},
(3) the upper and lower bounds \eqn{\mathbf{u} = (u_1, u_2, \dots, u_K)}
and \eqn{\boldsymbol{\ell} = (\ell_1, \ell_2, \dots, \ell_K)}.

The expected sample size is:
\deqn{
E[N \, | \, \delta]=\sum_{i=1}^K n_i P(\text{trial stops after analysis }i \, | \, \delta)
}

The probability that the trial stops after analysis $i$ is the sum of the
probabilities that the trial stops for efficacy after analysis $i$ and the
probability that the trial stops for futility after analysis $i$.

\deqn{
 E[N \, | \, \delta]=\sum_{i=1}^K n_i \Big[ P(\text{trial stops for efficacy after analysis }i \, | \, \delta) + \\
 P(\text{trial stops for futility after analysis }i \, | \, \delta) \Big]
}
}

\subsection{Variance of sample size}{

The definition of the variance for a random variable \eqn{X},

\deqn{
  Var(X) = E[X]^2 - E[X^2]
}

is used to calculate the variance for the expected sample size.
}

\subsection{Type I error}{

The type I error for the design (\eqn{\alpha}) is calculated by summing the
probabilities for stopping for efficacy under the null hypothesis for each
analysis.
}

\subsection{Power}{

The power for the design is calculated by summing the probabilities for
stopping for futility under the alternative to calculate the type II error
(\eqn{\beta}) and the subtracting this from 1.
}
}
\examples{
gsd_simulations <- function(n_analyses = 3,
                            upper_bounds = c(2.5, 2, 1.5),
                            lower_bounds = c(0, 0.75, 1.5),
                            n_patients = c(20, 40, 60),
                            null_hypothesis = 0,
                            alt_hypothesis = 0.5,
                            variance = 1)
}
