% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd_simulations.R
\name{gsd_simulations}
\alias{gsd_simulations}
\title{Title}
\usage{
gsd_simulations(
  n_analyses = 3,
  upper_bounds = c(2.5, 2, 1.5),
  lower_bounds = c(0, 0.75, 1.5),
  n_patients = c(20, 40, 60),
  null_hypothesis = 0,
  alt_hypothesis = 0.5,
  variance = 1
)
}
\arguments{
\item{variance}{}
}
\description{
Title
}
